Filter 적용 :전체 컨트롤러 수정, 수정 삭제 유저 검증 로직 추가

+++ community-BE/src/main/java/com/ktb/community_BE/service/CommentService.java
@@ -55,8 +55,10 @@ public class CommentService {
     //댓글 수정
     @Transactional
     public CommentDto updateComment(CommentDto commentRequest) {
-        Comment comment = commentRepository.findById(commentRequest.getId()).orElseThrow(() -> new IllegalArgumentException("Post not found"));
-
+        Comment comment = commentRepository.findById(commentRequest.getId()).orElseThrow(() -> new IllegalArgumentException("comment not found"));
+        if(commentRequest.getUserId() != comment.getUser().getId()){
+            throw new IllegalArgumentException("비정상적인 접근입니다.");
+        }
         if (commentRequest.getContent() != null){
             comment.changeContent(commentRequest.getContent());
         }
@@ -74,11 +76,13 @@ public class CommentService {
 
     //댓글 삭제
     @Transactional
-    public void deleteComment(Long id, Long postId){
+    public void deleteComment(Long id, Long postId, Long userId){
         Post post = postRepository.findById(postId).orElseThrow(() -> new IllegalArgumentException("Post not found"));
-
+        Comment comment = commentRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Post not found"));
+        if(userId != comment.getUser().getId()){
+            throw new IllegalArgumentException("비정상적인 접근입니다.");
+        }
         post.minusCommentCount();
-
         commentRepository.deleteById(id);
     }
